# Count perf optimizations.